Move ConditionPart struct to separate h/cpp file
